7.14
- added new tiles
- readded starting maps
- changed fireball texture
- changed fireball script with vector math
- changed many doubles to floats
- renamed scriptManager to Lua

+++ 5.24/Camera.h
@@ -10,9 +10,9 @@ public:
 	Camera();
 
 	void toggle();
-	void move(int dir, double dis = 0);
+	void move(int dir, float dis = 0);
 	void update();
-	void update(double x, double y);
+	void update(float x, float y);
 	void center(Entity *entity);  // change camera focus
 	void zoom(const float &amount);
 
@@ -23,8 +23,8 @@ public:
 
 	bool get_locked();
 
-	double get_x() { return _x; }
-	double get_y() { return _y; }
+	float get_x() { return _x; }
+	float get_y() { return _y; }
 
 	float get_scale()	  { return _uniform_scale;	  }
 	double get_rotation() { return _uniform_rotation; }
@@ -33,8 +33,8 @@ public:
 	friend class Renderer;
 private:
 	const char *_FILE_PATH = "Data/system.txt";
-	const double _SPEED = 500.0;
-	double _x, _y, _speed;
+	const float _SPEED = 500.0f;
+	float _x, _y, _speed;
 	bool _is_locked;
 	Entity *_entity;  // entity to center camera on
