+++ 5.24/InputHandler.h
@@ -7,6 +7,6 @@
 void move_entity(Entity *entity, int dir);
 void move_camera(int dir);
 void toggle_camera();
-void cast_spell(double x, double y);
+void cast_spell(float x, float y);
 
 #endif
